2474: move `assertConsistentCardInfo` to card_info_utils.dart since it does not only belong to the verification process, add _assertCardIsAlreadyValid function explicitly for static card verification
